+++ DL/UsersGetDTO.cs
@@ -14,7 +14,7 @@ namespace DL
         public string ApellidoMaterno { get; set; }
         public DateTime FechaNacimiento {  get; set; }
         public byte[]? Imagen {  get; set; }
-        public int IdRol {  get; set; }
+        public byte IdRol {  get; set; }
         public string NombreRol { get; set; }
     }
 }
